backend/app/main.py
from fastapi import FastAPI
from app.database import database,Base,engine
from app.routers import users,votes,stocks
import uvicorn
from starlette.middleware.cors import CORSMiddleware

app = FastAPI()
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"]
)
app.include_router(users.router)
app.include_router(votes.router)
app.include_router(stocks.router)

@app.on_event("startup")
async def startup():
    Base.metadata.create_all(engine)
    await database.connect()


@app.on_event("shutdown")
async def shutdown():
    await database.disconnect()


if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
