backend/requirements.txt
aiosqlite
appdirs
bcrypt
black
certifi
cffi
chardet
click
databases
Deprecated
dnspython
email-validator
fastapi
fastapi-users
finnhub-python
h11
httpcore
httpx
httpx-oauth
idna
makefun
mypy-extensions
passlib
pathspec
pycparser
pydantic
PyJWT
python-multipart
regex
requests
rfc3986
six
sniffio
SQLAlchemy
SQLAlchemy-Utils
starlette
toml
typed-ast
typing-extensions
urllib3
wrapt
